chore: refactor for payment e2e including viewport context aware
