+++ projects/storefrontapp-e2e-cypress/cypress/helpers/payment-methods.ts
@@ -1,9 +1,4 @@
-import {
-  addProductToCartViaAutoComplete,
-  addProductToCartViaSearchPage,
-  loginRegisteredUser,
-} from './cart';
-import { visitHomePage, waitForPage } from './checkout-flow';
+import { waitForPage } from './checkout-flow';
 
 interface PaymentDetail {
   accountHolderName: string;
@@ -25,7 +20,7 @@ interface PaymentDetail {
   };
 }
 
-const testPaymentDetail: PaymentDetail[] = [
+export const testPaymentDetail: PaymentDetail[] = [
   {
     accountHolderName: 'test user',
     cardNumber: 4111111111111111,
@@ -66,23 +61,6 @@ const testPaymentDetail: PaymentDetail[] = [
   },
 ];
 
-export function checkAnonymous() {
-  it('should redirect anonymous user to login page', () => {
-    cy.visit('/my-account/payment-details');
-    cy.location('pathname').should('contain', '/login');
-  });
-}
-
-export function verifyText() {
-  cy.get('cx-payment-methods').within(() => {
-    cy.get('.cx-payment .cx-header').should('contain', 'Payment methods');
-    cy.get('.cx-payment .cx-body').should(
-      'contain',
-      'New payment methods are added during checkout.'
-    );
-  });
-}
-
 export function verifyPaymentCard(cardLength: number) {
   cy.get('.cx-payment .cx-body').then(() => {
     cy.get('cx-card').should('exist');
@@ -90,102 +68,16 @@ export function verifyPaymentCard(cardLength: number) {
   });
 }
 
-export function setOtherPaymentToDefault() {
-  cy.get('cx-payment-methods')
-    .findByText('Set as default')
-    .click({ force: true });
-
-  const firstCard = cy.get('.cx-payment-card').first();
-  firstCard.should('contain', '✓ DEFAULT');
-  firstCard.should('contain', '1234');
-  firstCard.should('contain', `Expires: 03/2126`);
-}
-
-export function deletePayment() {
-  cy.findAllByText('Delete').first().click({ force: true });
-
-  // should see confirmation message
-  cy.get('.cx-card-delete-msg').should(
-    'contain',
-    'Are you sure you want to delete this payment method?'
-  );
-
-  cy.get('.btn-secondary').should('contain', 'Cancel');
-  cy.get('.btn-secondary').click({ force: true });
-  cy.get('.cx-card-body__delete-ms').should('not.exist');
-
-  // delete the payment
-  cy.findAllByText('Delete').first().click({ force: true });
-  cy.get('.btn-primary').should('contain', 'Delete');
-  cy.get('.btn-primary').click({ force: true });
-  cy.get('.cx-payment-card').should('have.length', 1);
-
-  // verify remaining address is now the default one
-  const defaultCard = cy.get('.cx-payment-card');
-  defaultCard.should('contain', '✓ DEFAULT');
-  defaultCard.should('contain', 'test user');
-}
-
-export function visitPaymentDetailsPage(isMobile: boolean = false) {
+export function visitPaymentDetailsPage() {
   const paymentDetailPage = waitForPage(
     '/my-account/payment-detail',
     'getPaymentDetail'
   );
-  cy.selectUserMenuOption({
-    option: 'Payment Details',
-    isMobile,
-  });
+  cy.selectUserMenuOption({ option: 'Payment Details' });
   cy.wait(`@${paymentDetailPage}`).its('status').should('eq', 200);
 }
 
-export function paymentMethodsTest(isMobile: boolean = false) {
-  it('should visit payment details page', () => {
-    loginRegisteredUser();
-    visitPaymentDetailsPage(isMobile);
-  });
-
-  it('should render page with no payment methods', () => {
-    verifyText();
-  });
-
-  it('should add product to cart via search', () => {
-    addProductToCartViaAutoComplete(false);
-  });
-
-  it('should create a payment detail using POST request', () => {
-    addPaymentMethod(testPaymentDetail[0]);
-  });
-
-  it('should render page with only one payment methods', () => {
-    visitHomePage();
-    visitPaymentDetailsPage(isMobile);
-    verifyPaymentCard(1);
-  });
-
-  it('should add a second product via search', () => {
-    addProductToCartViaSearchPage(false);
-  });
-
-  it('should add a second payment detail using POST request', () => {
-    addPaymentMethod(testPaymentDetail[1]);
-  });
-
-  it('should render page with only two payment methods', () => {
-    visitHomePage();
-    visitPaymentDetailsPage(isMobile);
-    verifyPaymentCard(2);
-  });
-
-  it('should set additional payment method as default', () => {
-    setOtherPaymentToDefault();
-  });
-
-  it('should be able to delete payment method', () => {
-    deletePayment();
-  });
-}
-
-function addPaymentMethod(paymentDetail: PaymentDetail) {
+export function addPaymentMethod(paymentDetail: PaymentDetail) {
   cy.get('.cx-total')
     .first()
     .then(($cart) => {
